contracts/requirements.txt
web3>=6.0.0
solcx>=0.2.0
python-dotenv>=1.0.0
